segregate ortools and cvx solvers
