+++ example_test/run_populationsim.py
@@ -12,8 +12,8 @@ from activitysim.core.config import handle_standard_args
 
 from populationsim import steps
 from populationsim.util import setting
-from populationsim import integerizer
-from populationsim import simul_integerizer
+from populationsim import lp
+from populationsim import multi_integerizer
 
 
 handle_standard_args()
@@ -35,8 +35,8 @@ logger.info("meta_control_data: %s"
 logger.info("control_file_name: %s"
             % setting('control_file_name'))
 
-logger.info("USE_CVXPY: %s" % integerizer.use_cvxpy())
-logger.info("USE_SIMUL_INTEGERIZER: %s" % simul_integerizer.use_simul_integerizer())
+logger.info("USE_CVXPY: %s" % lp.use_cvxpy())
+logger.info("USE_SIMUL_INTEGERIZER: %s" % multi_integerizer.use_simul_integerizer())
 
 
 # get the run list (name was possibly specified on the command line with the -m option)
