+++ setup.cfg
@@ -2,3 +2,7 @@
 max-line-length = 100
 exclude = ./calm_data_prep/convert_rsg_data.py,./sandbox/
 
+[coverage:run]
+omit =
+    sandbox/*
+    populationsim/lp_cvx.py
